backend -v1 correct apis , collection and automatic testing/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyodbc
passlib[bcrypt]
argon2-cffi
python-multipart
python-jose[cryptography]
python-dotenv
pytest
httpx